Make backup restoration more granular and friendly
ADOM can exit with rc 1 if you quit at the load screen, causing some things
to think it is a crash or abnormal exit spuriously.

Whether spurious or not, if ADOM exited abnormally, the backup and live save
folders will be compared.  Any missing saves will be prompted for, and also
any saves that exist in both places but are different.

It's not possible to lose more than one live save at once, but this code
handles that anyway.