Ignore non-.svg files in savedg folder
I don't currently see a need to do anything with *.bak files.  The server makes a backup on ADOM exit, and .bak files are created on game load. There may be a case for multi-character sessions.
But in the case of loading CharA, saving and returning to menu, starting CharB and crashing, the live version of CharA.svg is valid. The wrapper will prompt to replace it, but the user should read the date on the prompt and answer 'n'.
In the case of starting CharC, saving and returning to menu and reloading CharC, the server will not have a backup copy.  I think the solution to this is the autosaver.